requirements.txt
Flask
pymongo
Pillow
opencv-python
pytesseract
